Restructured folders hierachy
