+++ include/BeastEngine/Core/Helpers.h
