+++ include/BeastEngine/Core/Loggers/LoggersTypes.h
