+++ include/BeastEngine/Core/Loggers/StaticLogger.h
@@ -1,6 +1,6 @@
 #pragma once
-#include "BeastEngine/Definitions/Types.h"
-#include "BeastEngine/Definitions/Helpers.h"
+#include "BeastEngine/Core/Types.h"
+#include "BeastEngine/Core/Helpers.h"
 
 namespace be::internals
 {
