+++ include/BeastEngine/Core/Versions.h.in
