+++ tests/Core/Loggers/LoggersFactoryTest.h
